requirements.txt
streamlit
pandas
openpyxl
SpeechRecognition
streamlit-mic-recorder
edge-tts
google-generativeai